+++ lib/modules/forms/forms_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: forms
-// Updated at: 2026-06-24 16:37:55
-// Commit: refactor(forms): extract reusable atomic widgets from forms
+// Updated at: 2026-06-24 18:12:12
+// Commit: fix(forms): fix memory leak by cancelling forms subscription
 
 class formsComponentState {
-  final String commitRef = '2266';
-  final String lastMessage = 'refactor(forms): extract reusable atomic widgets from forms';
-  final DateTime timestamp = DateTime.parse('2026-06-24T16:37:55+05:00');
+  final String commitRef = '2272';
+  final String lastMessage = 'fix(forms): fix memory leak by cancelling forms subscription';
+  final DateTime timestamp = DateTime.parse('2026-06-24T18:12:12+05:00');
 }